Move SoftdeviceExternalController to ble.rs
